Add auth middleware to endpoints

+++ routes/web.php
@@ -17,21 +17,21 @@ use Illuminate\Support\Facades\Route;
 
 // Listings CRUD
 Route::get('/', [ListingController::class, 'index']);
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 Route::post('/listings', [ListingController::class, 'store']);
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 // Users CRUD
-Route::get("/register", [UserController::class, 'create']);
-Route::post("/users", [UserController::class, 'store']);
+Route::get("/register", [UserController::class, 'create'])->middleware('guest');
+Route::post("/users", [UserController::class, 'store'])->middleware('guest');
 
 //authentication
-Route::get("/login", [UserController::class, 'login']);
+Route::get("/login", [UserController::class, 'login'])->name('login')->middleware('guest');
 Route::post("/users/authenticate", [UserController::class, 'authenticate']);
-Route::POST("/logout", [UserController::class, 'logout']);
+Route::POST("/logout", [UserController::class, 'logout'])->middleware('auth');
